Added loging via slf4j

+++ ContactListWeb/src/main/java/edu/itechart/contactlist/commands/DeleteContactCommand.java
@@ -2,6 +2,7 @@ package edu.itechart.contactlist.commands;
 
 import edu.itechart.contactlist.service.ContactService;
 import edu.itechart.contactlist.service.ServiceException;
+import org.apache.commons.lang3.StringUtils;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -14,8 +15,10 @@ public class DeleteContactCommand implements Command {
     public String execute(HttpServletRequest request, HttpServletResponse response) throws CommandException {
         try {
             String[] items = request.getParameterValues(REQUEST_PARAM_NAME);
-            for (String item : items) {
-                ContactService.delete(Long.parseLong(item));
+            if (items != null) {
+                for (String item : items) {
+                    ContactService.delete(Long.parseLong(item));
+                }
             }
         } catch (ServiceException e) {
             throw new CommandException(e);
